Moved Command class to Commands module
