+++ lib/branch_io_cli.rb
@@ -1,5 +1,4 @@
 require "branch_io_cli/cli"
-require "branch_io_cli/command"
 require "branch_io_cli/commands"
 require "branch_io_cli/helper"
 require "branch_io_cli/version"
