+++ lib/branch_io_cli/commands.rb
@@ -1,2 +1,3 @@
+require "branch_io_cli/commands/command"
 require "branch_io_cli/commands/setup_command"
 require "branch_io_cli/commands/validate_command"
