+++ lib/branch_io_cli/commands/setup_command.rb
@@ -14,7 +14,7 @@ module BranchIOCLI
         update_podfile(options) || update_cartfile(options, xcodeproj)
 
         target_name = options.target # may be nil
-        is_app_target = !Helper::ConfigurationHelper.target.extension_target_type?
+        is_app_target = config_helper.target.extension_target_type?
 
         if is_app_target && options.validate &&
            !helper.validate_team_and_bundle_ids_from_aasa_files(xcodeproj, target_name, @domains)
@@ -97,7 +97,7 @@ module BranchIOCLI
         # 4. Add to target dependencies
         frameworks_group = project.frameworks_group
         branch_framework = frameworks_group.new_file "Carthage/Build/iOS/Branch.framework"
-        target = Helper::ConfigurationHelper.target
+        target = config_helper.target
         target.frameworks_build_phase.add_file_reference branch_framework
 
         # 5. Add to copy-frameworks build phase
